iris_service/requirements.txt
fastapi
numpy==1.16.4
scikit-learn==0.20.1
uvicorn==0.9.0
prometheus-client
starlette_prometheus
python-logging-loki